+++ IliaCodeChallenge/IliaCodeChallenge/View/Scenes/Home/MovieCollectionViewCell.swift
@@ -1,0 +1,17 @@
+//
+//  MovieCollectionViewCell.swift
+//  IliaCodeChallenge
+//
+//  Created by Stephanie Torisu on 27/01/22.
+//
+
+import UIKit
+
+class MovieCollectionViewCell: UICollectionViewCell {
+
+    override func awakeFromNib() {
+        super.awakeFromNib()
+        // Initialization code
+    }
+
+}
