+++ IliaCodeChallenge/IliaCodeChallenge/View/Scenes/MovieDetail/MovieDetailViewController.swift
@@ -1,0 +1,29 @@
+//
+//  MovieDetailViewController.swift
+//  IliaCodeChallenge
+//
+//  Created by Stephanie Torisu on 27/01/22.
+//
+
+import UIKit
+
+class MovieDetailViewController: ViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Do any additional setup after loading the view.
+    }
+
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destination.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
